break files up
