+++ src/gameinput.js
@@ -3,9 +3,14 @@
  * @source: https://raw.githubusercontent.com/lunarcloud/gameinputjs/master/gameinput.js
  * @license magnet:?xt=urn:btih:d3d9a9a6595521f9666a5e94cc830dab83b65699&dn=expat.txt MIT (Expat) License
  */
-import { Vector2, GamepadAPI, GameInputModel, GameInputSchemaType, SchemaAxisButton } from './gameinput.types.js'
-import { GameInputModels } from './gameinput.models.js'
-
+import {GamepadAPI} from './gamepad-api.js'
+import {GameInputModel} from './gameinput-model.js'
+import {GameInputSchema} from './gameinput-schema.js'
+import {SchemaAxisButton} from './schema-axis-button.js'
+import {GameInputModels} from './gameinput.models.js'
+import {GameInputPlayer} from './gameinput-player.js'
+import {GamepadSchemaNames} from './gamepad-schema-names.js'
+import {Vector2} from './vector2.js'
 
 const osStrings = [
     {s:'Android', r:/Android/},
@@ -13,276 +18,33 @@ const osStrings = [
     {s:'Windows', r:/Windows/},
     {s:'macOS', r:/Mac/},
     {s:'Linux', r:/(Linux|X11)/}
-];
-
-function normalize(val, min, max) {
-    return (val-min)/(max-min);
-}
-
-class GameInputPlayer {
-
-    #gameInput
-
-    constructor(gameInput, number) {
-        this.#gameInput = gameInput
-        this.number = number;
-        this.index = number - 1;
-
-        this.type;
-        this.model;
-        this.schema;
-        this.theme;
-        this.state;
-        this.analog;
-
-        this.previous = {
-            type: undefined,
-            model: undefined,
-            schema: undefined,
-            state: undefined,
-            analog: undefined,
-        };
-
-        this.buttonDownActions = {};
-        this.buttonUpActions = {};
-
-        for (var i in GameInput.SchemaNames) {
-            this.buttonDownActions[GameInput.SchemaNames[i]] = [];
-            this.buttonUpActions[GameInput.SchemaNames[i]] = [];
-        }
-    }
-    buttonDown(schemaName) {
-        this.#gameInput.buttonDown(this.index, schemaName);
-        for (var action in this.buttonDownActions[schemaName])
-            this.buttonDownActions[schemaName][action]();
-    }
-    buttonUp(schemaName) {
-        this.#gameInput.buttonUp(this.index, schemaName);
-        for (var action in this.buttonUpActions[schemaName])
-            this.buttonUpActions[schemaName][action]();
-    }
-    onButtonDown(schemaName, action) {
-        if (schemaName in GameInput.SchemaNames === false) throw "Must be SchemaNames";
-        if (typeof (action) !== "function") throw "Action must be a function";
-
-        this.buttonDownActions[schemaName].push(action);
-    }
-    onButtonUp(schemaName, action) {
-        if (schemaName in GameInput.SchemaNames === false) throw "Must be SchemaNames";
-        if (typeof (action) !== "function") throw "Action must be a function";
-
-        this.buttonUpActions[schemaName].push(action);
-    }
-    getStickVector(stick) {
-        if (stick != "l" && stick != "r") throw "Must be l or r";
-
-        var x = 0;
-        var y = 0;
-
-        if (this.schema[stick + "_up"] instanceof SchemaAxisButton) {
-            if (this.schema[stick + "_up"].direction == "negative") {
-                y -= this.analog[stick + "_up"] < this.schema[stick + "_up"].deadZone ? Math.abs(this.analog[stick + "_up"]) : 0;
-            } else {
-                y -= this.analog[stick + "_up"] > this.schema[stick + "_up"].deadZone ? Math.abs(this.analog[stick + "_up"]) : 0;
-            }
-        } else {
-            y -= 0.7;
-        }
-
-        if (this.schema[stick + "_down"] instanceof SchemaAxisButton) {
-            if (this.schema[stick + "_down"].direction == "negative") {
-                y += this.analog[stick + "_down"] < this.schema[stick + "_down"].deadZone ? Math.abs(this.analog[stick + "_down"]) : 0;
-            } else {
-                y += this.analog[stick + "_down"] > this.schema[stick + "_down"].deadZone ? Math.abs(this.analog[stick + "_down"]) : 0;
-            }
-        } else {
-            x += 0.7;
-        }
-
-        if (this.schema[stick + "_left"] instanceof SchemaAxisButton) {
-            if (this.schema[stick + "_left"].direction == "negative") {
-                x -= this.analog[stick + "_left"] < this.schema[stick + "_left"].deadZone ? Math.abs(this.analog[stick + "_left"]) : 0;
-            } else {
-                x -= this.analog[stick + "_left"] > this.schema[stick + "_left"].deadZone ? Math.abs(this.analog[stick + "_left"]) : 0;
-            }
-        } else {
-            x -= 0.7;
-        }
-
-        if (this.schema[stick + "_right"] instanceof SchemaAxisButton) {
-            if (this.schema[stick + "_right"].direction == "negative") {
-                x += this.analog[stick + "_right"] < this.schema[stick + "_right"].deadZone ? Math.abs(this.analog[stick + "_right"]) : 0;
-            } else {
-                x += this.analog[stick + "_right"] > this.schema[stick + "_right"].deadZone ? Math.abs(this.analog[stick + "_right"]) : 0;
-            }
-        } else {
-            x += 0.7;
-        }
-
-        return new Vector2(x, y);
-
-    }
-    getNormalizedStickVector(stick) {
-        var stickInput = this.getStickVector(stick);
-        var radialDeadZone = 0;
-
-        for (var direction in ["up", "down", "left", "right"]) {
-            if (this.schema[stick + "_" + direction] instanceof SchemaAxisButton) {
-                if (this.schema[stick + "_" + direction].deadZone > radialDeadZone) {
-                    radialDeadZone = this.schema[stick + "_" + direction].deadZone;
-                }
-            }
-        }
-
-        if (stickInput.magnitude < radialDeadZone) {
-            return new Vector2(0, 0);
-        } else {
-            return stickInput.normalize().scale((stickInput.magnitude() - radialDeadZone) / (1 - radialDeadZone));
-        }
-    }
-    getNormalizedTriggerValue(trigger) {
-        if (trigger != "l" && trigger != "r") throw "Must be l or r";
-        trigger += "_trigger";
-
-        if (this.schema[trigger] instanceof SchemaKey
-            || this.schema[trigger] instanceof SchemaButton) {
-            return this.state[trigger] ? 1 : 0;
-        }
-        // else  this.schema[trigger] instanceof SchemaAxisButton
-        return normalize(
-        /*val*/ this.state[trigger],
-            /*min*/ this.schema[trigger].deadZone,
-            /*max*/ 1
-        );
-    }
-    /**
-     * @desc    Gets the button text
-     * @param   schemaName      name of the button or axisValue
-     * @param   symbolsAsWords  whether or not to convert Ragdoll's "x □ o △" to "cross square circle triangle"
-     */
-    getButtonText(schemaName, symbolsAsWords) {
-        if (this.model && this.model.type) {
-            var text = this.model.type.schemaNames[schemaName];
-
-            if (symbolsAsWords !== true) return text;
-
-            switch (text) {
-                case "▶":
-                    return "start";
-                case "x":
-                    return "cross";
-                case "o":
-                    return "circle";
-                case "□":
-                    return "square";
-                case "△":
-                    return "triangle";
-                default:
-                    return text;
-            }
-        }
-    }
-};
-
+]
 
 /**
  * GameInput
  * @brief   Game Input System
  * @desc    System for using a gamepad control scheme for games
  */
-class GameInput {
+export default class GameInput {
 
-    debug = true; // disable to remove console output
-
-    /* Helper function */
-    static toASCII (text) {
-        return text.replace(/[^\x00-\x7F]/g, ""); /* eslint-disable-line no-control-regex */
-    }
-
+    /**
+     * Current OS.
+     * @type {import('./gameinput-model.js').OSName}
+     */
     static os = "Other";
+
+    /**
+     * Current Browser.
+     * @type {'Chrome'|'Firefox'|'Other'} 
+     */
     static browser = "Other";
-    static firstPress = false;
 
-    static canUseGamepadAPI ()
-    {
-        return "getGamepads" in navigator;
-    }
-
-    buttonDownActions = [];
-    buttonUpActions = [];
-
-    onButtonDown (action) {
-        if (typeof(action) !== "function") throw "Action must be a function";
-        this.buttonDownActions.push(action);
-    }
-
-    onButtonUp (action) {
-        if (typeof(action) !== "function") throw "Action must be a function";
-        this.buttonUpActions.push(action);
-    }
-
-    buttonDown (player, schemaName) {
-        for ( var action in this.buttonDownActions) {
-            if (typeof(this.buttonDownActions[action]) === "function")
-                this.buttonDownActions[action](player, schemaName);
-        }
-    }
-
-    buttonUp (player, schemaName) {
-        for ( var action in this.buttonUpActions) {
-            if (typeof(this.buttonUpActions[action]) === "function")
-                this.buttonUpActions[action](player, schemaName);
-        }
-    }
-
-    Players = [
-        new GameInputPlayer(this, 1),
-        new GameInputPlayer(this, 2),
-        new GameInputPlayer(this, 3),
-        new GameInputPlayer(this, 4)
-    ];
-
-    Connection = {
-        GamePadMapping: {
-            0 : 0,
-            1 : 1,
-            2 : 2,
-            3 : 3
-        },
-        Gamepads: [undefined, undefined, undefined, undefined]
-    }
-
-    getPlayer (index) {
-        return this.Players[this.Connection.GamePadMapping[index]];
-    }
-
-    static SchemaNames = {
-        d_up        :   "d_up",
-        d_down      :   "d_down",
-        d_left      :   "d_left",
-        d_right     :   "d_right",
-        menu        :   "menu",
-        button0     :   "button0",
-        button1     :   "button1",
-        button2     :   "button2",
-        button3     :   "button3",
-        l_up        :   "l_up",
-        l_down      :   "l_down",
-        l_left      :   "l_left",
-        l_right     :   "l_right",
-        r_up        :   "r_up",
-        r_down      :   "r_down",
-        r_left      :   "r_left",
-        r_right     :   "r_right",
-        l_button    :   "l_button",
-        r_button    :   "r_button",
-        l_trigger   :   "l_trigger",
-        r_trigger   :   "r_trigger"
-    }
-
+    /**
+     * Gamepad Models.
+     */
     static Models = {
         UnknownStandardMapping: new GameInputModel(
-            GameInputSchemaType.Hedgehog,
+            GameInputSchema.Hedgehog,
             "generic",
             undefined,
             undefined,
@@ -290,28 +52,28 @@ class GameInput {
 
         Generic: [
             new GameInputModel(
-                GameInputSchemaType.Hedgehog,
+                GameInputSchema.Hedgehog,
                 "xbox360",
                 "XInput",
                 undefined,
                 GamepadAPI.Stardard
             ),
             new GameInputModel(
-                GameInputSchemaType.Hedgehog,
+                GameInputSchema.Hedgehog,
                 "xbox360",
                 "xinput",
                 undefined,
                 GamepadAPI.Stardard
             ),
             new GameInputModel(
-                GameInputSchemaType.Hedgehog,
+                GameInputSchema.Hedgehog,
                 "xbox360",
                 "XBox 360",
                 undefined,
                 GamepadAPI.Stardard
             ),
             new GameInputModel(
-                GameInputSchemaType.Hedgehog,
+                GameInputSchema.Hedgehog,
                 "generic",
                 "Logitech Rumblepad 2",
                 undefined,
@@ -333,7 +95,7 @@ class GameInput {
                     // TODO r_trigger
             )),
             new GameInputModel(
-                GameInputSchemaType.Hedgehog,
+                GameInputSchema.Hedgehog,
                 "generic",
                 "Logitech Dual Action",
                 undefined,
@@ -355,7 +117,7 @@ class GameInput {
                     // TODO r_trigger
             )),
             new GameInputModel(
-                GameInputSchemaType.Hedgehog,
+                GameInputSchema.Hedgehog,
                 "generic",
                 "STANDARD GAMEPAD",
                 undefined,
@@ -366,28 +128,210 @@ class GameInput {
         Specific: GameInputModels // imported
     }
 
-    loopingUpdate = true;
+    static toASCII (text) {
+        return text.replace(/[^\x00-\x7F]/g, ""); /* eslint-disable-line no-control-regex */
+    }
 
+    static canUseGamepadAPI ()
+    {
+        return "getGamepads" in navigator;
+    }
+    
+    Players = [
+        new GameInputPlayer(this, 1),
+        new GameInputPlayer(this, 2),
+        new GameInputPlayer(this, 3),
+        new GameInputPlayer(this, 4)
+    ];
+
+    Connection = {
+        GamePadMapping: {
+            0 : 0,
+            1 : 1,
+            2 : 2,
+            3 : 3
+        },
+        Gamepads: [undefined, undefined, undefined, undefined]
+    }
+    /**
+     * Whether to display debug logs.
+     * @type {boolean}
+     */
+    debug = true;
+
+    /**
+     * Whether are actively updating.
+     * @type {boolean}
+     */
+    #updateActive = true;
+
+    /**
+     * Whether we've received the first button press.
+     */
+    firstPress = false;
+
+    /**
+     * Actions to perform on button down.
+     * @type {Array<Function>}
+     */
+    buttonDownActions = [];
+    
+    /**
+     * Actions to perform on button up.
+     * @type {Array<Function>}
+     */
+    buttonUpActions = [];
+    
+    /**
+     * Last known gamepads count.
+     * @type {number}
+     */
+    lastCheckedNumberOfGamepads = -1;
+    
+    /**
+     * Actions to perform after players reshuffled.
+     * @type {Array<Function>}
+     */
+    reshufflePlayersActions = [];
+
+    /**
+     * Constructor.
+     */
+    constructor () {
+        /* Detect OS */
+        for (var id in osStrings) {
+            if (osStrings[id].r.test(navigator.userAgent)) { this.os = osStrings[id].s; break; }
+        }
+
+        /* Detect Browser */
+        if (/Chrome/.test(navigator.userAgent)) {
+            this.browser = "Chrome";
+        } else if (/Firefox/.test(navigator.userAgent)) {
+            this.browser = "Firefox";
+        } // else { this.browser = "Other" }
+
+
+        this.startUpdateLoop();
+
+        // Start watching for gamepads joining and leaving
+        if (GameInput.canUseGamepadAPI())
+        {
+            if (this.debug)
+                console.debug("Gamepad connection loop beginning")
+            this.connectionWatchLoop();
+
+            // warning, these are very unreliable!
+            window.addEventListener("gamepadconnected", function(e) {
+                if (this.debug)
+                    console.debug("Gamepad connected at index %d: %s. %d buttons, %d axes.",
+                        e.gamepad.index, e.gamepad.id,
+                        e.gamepad.buttons.length, e.gamepad.axes.length);
+            }, false);
+
+            window.addEventListener("gamepaddisconnected", function(e) {
+                if (this.debug)
+                    console.debug("Gamepad disconnected from index %d: %s",
+                        e.gamepad.index, e.gamepad.id);
+            }, false);
+        }
+    }
+
+    /**
+     * Add an action to "reshuffle players" events.
+     * @param {Function} action Action to add.
+     */
+    onReshufflePlayers (action)
+    {
+        if (typeof(action) !== "function") 
+            throw "Action must be a function";
+        this.reshufflePlayersActions.push(action);
+    }
+
+    /**
+     * Add an action to "button down" events.
+     * @param {Function} action Action to add.
+     */
+    onButtonDown (action) {
+        if (typeof(action) !== "function") 
+            throw "Action must be a function";
+        this.buttonDownActions.push(action);
+    }
+
+    /**
+     * Add an action to "button down" events.
+     * @param {Function} action Action to add.
+     */
+    onButtonUp (action) {
+        if (typeof(action) !== "function") 
+            throw "Action must be a function";
+        this.buttonUpActions.push(action);
+    }
+
+    /**
+     * Add an action to "button down" events for a particular player.
+     * @param {number} player   Player to add action for.
+     * @param {Function} action Action to add.
+     */
+    buttonDown (player, schemaName) {
+        for ( var action in this.buttonDownActions) {
+            if (typeof(this.buttonDownActions[action]) === "function")
+                this.buttonDownActions[action](player, schemaName);
+        }
+    }
+
+    /**
+     * Add an action to "button up" events for a particular player.
+     * @param {number} player   Player to add action for.
+     * @param {Function} action Action to add.
+     */
+    buttonUp (player, schemaName) {
+        for ( var action in this.buttonUpActions) {
+            if (typeof(this.buttonUpActions[action]) === "function")
+                this.buttonUpActions[action](player, schemaName);
+        }
+    }
+
+    /**
+     * Get player for gamepad number.
+     * @param {0|1|2|3} index gamepad index
+     * @returns {GameInputPlayer} Player
+     */
+    getPlayer (index) {
+        return this.Players[this.Connection.GamePadMapping[index]];
+    }
+
+    /**
+     * Start the update loop.
+     */
     startUpdateLoop ()
     {
-        this.loopingUpdate = true;
-        this.nextUpdateLoop();
+        this.#updateActive = true;
+        this.#nextUpdateLoop();
     };
 
+    /**
+     * Stop the update loop.
+     */
     stopUpdateLoop ()
     {
-        this.loopingUpdate = false;
+        this.#updateActive = false;
     };
 
-    nextUpdateLoop ()
+    /**
+     * Perform the next update and loop.
+     */
+    #nextUpdateLoop ()
     {
-        if (this.loopingUpdate === false)
+        if (this.#updateActive === false)
             return;
 
         this.update();
-        requestAnimationFrame(() => this.nextUpdateLoop()); // way too slow!
-    };
+        requestAnimationFrame(() => this.#nextUpdateLoop()); // way too slow!
+    }
 
+    /**
+     * Gamepad state update.
+     */
     update ()
     {
         if (!GameInput.canUseGamepadAPI())
@@ -455,22 +399,20 @@ class GameInput {
         }
     }
 
-    gamepadsCount ()
+    /**
+     * Get the number of gamepads.
+     * @returns {number} count of gamepads
+     */
+    static gamepadsCount ()
     {
-        var gamepads = navigator.getGamepads();
-        var count = 0;
-
-        for (let i = 0; i < 4; i++) {
-            count += typeof gamepads[i] === "object" && gamepads[i] !== null ? 1 : 0;
-        }
-
-        return count;
+        return navigator.getGamepads().filter(Boolean).length;
     }
 
-    lastCheckedNumberOfGamepads = -1;
-
+    /**
+     * Loop that looks for gamepad updates
+     */
     connectionWatchLoop () {
-        var currentNumberOfGamepads = this.gamepadsCount();
+        var currentNumberOfGamepads = GameInput.gamepadsCount();
 
         if ( this.lastCheckedNumberOfGamepads !== currentNumberOfGamepads) {
             if ( this.debug )
@@ -483,6 +425,9 @@ class GameInput {
         requestAnimationFrame(() => this.connectionWatchLoop());
     }
 
+    /**
+     * Setup gamepads info.
+     */
     initialGamePadSetup ()
     {
         //clear gamepad information
@@ -593,53 +538,12 @@ class GameInput {
         }
     }
 
-    reshufflePlayersActions = [];
-
-    onReshufflePlayers (action)
-    {
-        this.reshufflePlayersActions.push(action);
-    }
-
-    constructor () {
-        /* Detect OS */
-        for (var id in osStrings) {
-            if (osStrings[id].r.test(navigator.userAgent)) { this.os = osStrings[id].s; break; }
-        }
-
-        /* Detect Browser */
-        if (/Chrome/.test(navigator.userAgent)) {
-            this.browser = "Chrome";
-        } else if (/Firefox/.test(navigator.userAgent)) {
-            this.browser = "Firefox";
-        } // else { this.browser = "Other" }
-
-
-        this.startUpdateLoop();
-
-        // Start watching for gamepads joining and leaving
-        if (GameInput.canUseGamepadAPI())
-        {
-            this.connectionWatchLoop();
-
-            // warning, these are very unreliable!
-            window.addEventListener("gamepadconnected", function(e) {
-                if (this.debug)
-                    console.debug("Gamepad connected at index %d: %s. %d buttons, %d axes.",
-                        e.gamepad.index, e.gamepad.id,
-                        e.gamepad.buttons.length, e.gamepad.axes.length);
-            }, false);
-
-            window.addEventListener("gamepaddisconnected", function(e) {
-                if (this.debug)
-                    console.debug("Gamepad disconnected from index %d: %s",
-                        e.gamepad.index, e.gamepad.id);
-            }, false);
-        }
-    }
 }
 
-export default GameInput
-export { GameInput, Vector2 }
+/**
+ * Export everything but GameInputModels (because it's inside GameInput already)
+ */
+export {GameInput, GamepadAPI, GameInputModel, GameInputSchema, SchemaAxisButton, GameInputPlayer, GamepadSchemaNames, Vector2}
 
 /**
  * @preserve
