+++ src/gameinput.models.js
@@ -3,7 +3,10 @@
  * @source: https://raw.githubusercontent.com/lunarcloud/gameinputjs/master/gameinput.models.js
  * @license magnet:?xt=urn:btih:d3d9a9a6595521f9666a5e94cc830dab83b65699&dn=expat.txt MIT (Expat) License
  */
-import { GameInputModel, GamepadAPI, GameInputSchemaType, SchemaAxisButton } from './gameinput.types.js'
+import {GameInputModel} from './gameinput-model.js'
+import {GameInputSchema} from './gameinput-schema.js'
+import {GamepadAPI} from './gamepad-api.js'
+import {SchemaAxisButton} from './schema-axis-button.js'
 
 const frenchVRStickDeadzone = 0.6;
 const frenchVRStickThreshold = 0.8;
@@ -39,13 +42,13 @@ const GravisGamepadProSchema = new GamepadAPI(
 
 const GameInputModels = [
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "xbox360",
         "Xbox 360 Controller (XInput STANDARD GAMEPAD)",
         "Windows"
     ),
     new GameInputModel(
-        GameInputSchemaType.Plumber,
+        GameInputSchema.Plumber,
         "nintendo-generic",
         "0925-8866-SFC/USB Pad",
         "Windows",
@@ -67,7 +70,7 @@ const GameInputModels = [
             5, 6, undefined, undefined
     )),
     new GameInputModel(
-        GameInputSchemaType.Plumber,
+        GameInputSchema.Plumber,
         "nintendo-generic",
         "SFC/USB Pad (Vendor: 0925 Product: 8866)",
         "Windows",
@@ -89,7 +92,7 @@ const GameInputModels = [
             5, 6, undefined, undefined
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "xbox360",
         "Xbox 360 Wireless Receiver (Vendor: 0000 Product: 0000)",
         "Linux",
@@ -108,7 +111,7 @@ const GameInputModels = [
             5,6,7,8
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "xbox360",
         "0000-0000-Xbox 360 Wireless Receiver",
         "Linux",
@@ -127,19 +130,19 @@ const GameInputModels = [
             5,6,7,8
     )),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll4,
+        GameInputSchema.Ragdoll4,
         "ds4",
         "Wireless Controller (STANDARD GAMEPAD Vendor: 054c Product: 05c4)",
         "Windows"
     ),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll4,
+        GameInputSchema.Ragdoll4,
         "ds4",
         "Wireless Controller (STANDARD GAMEPAD Vendor: 054c Product: 05c4)",
         "macOS"
     ),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll4,
+        GameInputSchema.Ragdoll4,
         "ds4",
         "054c-05c4-Wireless Controller",
         "Windows",
@@ -158,20 +161,19 @@ const GameInputModels = [
             5, 6, 7, 8
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "xbox360",
         "©Microsoft Corporation Controller (STANDARD GAMEPAD Vendor: 045e Product: 028e)",
         "Linux"
     ),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "xbox360",
         "Performance Designed Products Afterglow Gamepad for Xbox 360 (STANDARD GAMEPAD Vendor: 0e6f Product: 0213)",
-        "Linux",
-        GamepadAPI.Stardard
+        "Linux"
     ),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "xbox360",
         "Performance Designed Products Afterglow Gamepad for Xbox 360 (Vendor: 0e6f Product: 0213)",
         "Linux",
@@ -196,7 +198,7 @@ const GameInputModels = [
             new SchemaAxisButton(6)
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "xbox360",
         "0e6f-0213-Afterglow Gamepad for Xbox 360",
         "Linux",
@@ -221,7 +223,7 @@ const GameInputModels = [
             new SchemaAxisButton(6)
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "dc",
         "HuiJia  USB GamePad (Vendor: 0e8f Product: 3013)",
         "Linux",
@@ -240,7 +242,7 @@ const GameInputModels = [
             undefined, undefined, 5, 6
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "dc",
         "USB GamePad (Vendor: 0e8f Product: 3013)",
         "Windows",
@@ -259,7 +261,7 @@ const GameInputModels = [
             undefined, undefined, 5, 6
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "dc",
         "USB GamePad (Vendor: 0e8f Product: 3013)",
         "macOS",
@@ -278,13 +280,13 @@ const GameInputModels = [
         undefined, undefined, 5, 6
     )),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll,
+        GameInputSchema.Ragdoll,
         "ds3",
         "Sony PLAYSTATION(R)3 Controller (STANDARD GAMEPAD Vendor: 054c Product: 0268)",
         "Linux"
     ),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll,
+        GameInputSchema.Ragdoll,
         "ds3",
         "054c-0268-Sony PLAYSTATION(R)3 Controller",
         "Linux",
@@ -303,13 +305,13 @@ const GameInputModels = [
             11,12,9,10
     )),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll4,
+        GameInputSchema.Ragdoll4,
         "ds4",
         "Sony Computer Entertainment Wireless Controller (STANDARD GAMEPAD Vendor: 054c Product: 05c4)",
         "Linux"
     ),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll4,
+        GameInputSchema.Ragdoll4,
         "ds4",
         "054c-05c4-Sony Computer Entertainment Wireless Controller",
         "Linux",
@@ -331,7 +333,7 @@ const GameInputModels = [
             5, 6, 7, 8
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "joystick",
         "Mayflash Arcade Stick (Vendor: 0e8f Product: 0003)",
         "Windows",
@@ -353,7 +355,7 @@ const GameInputModels = [
             5, 6, 7, 8
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "joystick",
         "PLAYSTATION(R)3 Controller (STANDARD GAMEPAD Vendor: 054c Product: 0268)", /* Mayflash Arcade Stick */
         "macOS",
@@ -372,7 +374,7 @@ const GameInputModels = [
         5, 3, 4, 6
     )),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll,
+        GameInputSchema.Ragdoll,
         "joystick",
         "MY-POWER CO.,LTD. Mayflash Arcade Stick (STANDARD GAMEPAD Vendor: 0e8f Product: 0003)",
         "Linux",
@@ -394,7 +396,7 @@ const GameInputModels = [
             7,8,5,6
     )),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll,
+        GameInputSchema.Ragdoll,
         "joystick",
         "0e8f-0003-MY-POWER CO.,LTD. Mayflash Arcade Stick",
         "Linux",
@@ -416,7 +418,7 @@ const GameInputModels = [
             5,6,7,8
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "Logitech Logitech Dual Action (Vendor: 046d Product: c216)", // DirectInput (XInput uses 'standard' mapping)
         "Linux",
@@ -444,7 +446,7 @@ const GameInputModels = [
             /* r_trigger */  8
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "Logitech Gamepad F310 (Vendor: 046d Product: c21d)",
         "Linux",
@@ -472,7 +474,7 @@ const GameInputModels = [
             /* r_trigger */  new SchemaAxisButton(6)
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "046d-c21d-Logitech Gamepad F310", // XInput
         "Linux",
@@ -500,7 +502,7 @@ const GameInputModels = [
             /* r_trigger */  new SchemaAxisButton(6)
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "046d-c216-Logitech Logitech Dual Action", // DirectInput (XInput uses 'standard' mapping)
         "Linux",
@@ -528,7 +530,7 @@ const GameInputModels = [
             /* r_trigger */  8
     )),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll,
+        GameInputSchema.Ragdoll,
         "joystick",
         "Logitech Logitech Extreme 3D Pro (Vendor: 046d Product: c215)",
         "Linux",
@@ -550,7 +552,7 @@ const GameInputModels = [
             5,6,9,11
     )),
     new GameInputModel(
-        GameInputSchemaType.Ragdoll,
+        GameInputSchema.Ragdoll,
         "joystick",
         "046d-c215-Logitech Logitech Extreme 3D Pro",
         "Linux",
@@ -572,7 +574,7 @@ const GameInputModels = [
             5,6,9,11
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "Microsoft® Microsoft® SideWinder® Game Pad USB (Vendor: 045e Product: 0007)",
         "Linux",
@@ -600,7 +602,7 @@ const GameInputModels = [
     /* r_trigger */  8
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "045e-0007-Microsoft® Microsoft® SideWinder® Game Pad USB",
         "Linux",
@@ -628,7 +630,7 @@ const GameInputModels = [
     /* r_trigger */  8
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "ASUS Gamepad",
         "Android",
@@ -650,7 +652,7 @@ const GameInputModels = [
             7, 8, 5, 6
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "NYKO PLAYPAD PRO",
         "Android",
@@ -672,7 +674,7 @@ const GameInputModels = [
             5, 6, 7, 8
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "Moga 2 HID",
         "Android",
@@ -696,7 +698,7 @@ const GameInputModels = [
             undefined
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "Moga 2 HID (Vendor: 20d6 Product: 89e5)",
         "Linux",
@@ -720,7 +722,7 @@ const GameInputModels = [
             new SchemaAxisButton(5)
     )),
     new GameInputModel(
-        GameInputSchemaType.Plumber,
+        GameInputSchema.Plumber,
         "nintendo-generic",
         "(null) usb gamepad            (Vendor: 0810 Product: e501)",
         "Linux",
@@ -748,7 +750,7 @@ const GameInputModels = [
             /* r_trigger */  undefined
     )),
     new GameInputModel(
-    GameInputSchemaType.Plumber,
+    GameInputSchema.Plumber,
         "nintendo-generic",
         "0810-e501-usb gamepad           ",
         "Linux",
@@ -776,7 +778,7 @@ const GameInputModels = [
             /* r_trigger */  undefined
     )),
     new GameInputModel(
-        GameInputSchemaType.Plumber,
+        GameInputSchema.Plumber,
         "nintendo-generic",
         "usb gamepad            (Vendor: 0810 Product: e501)",
         "macOS",
@@ -804,7 +806,7 @@ const GameInputModels = [
             /* r_trigger */  undefined
     )),
     new GameInputModel(
-    GameInputSchemaType.Plumber,
+    GameInputSchema.Plumber,
         "nintendo-generic",
         "810-e501-usb gamepad           ",
         "macOS",
@@ -832,38 +834,37 @@ const GameInputModels = [
             /* r_trigger */  undefined
     )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "Gravis GamePad Pro USB  (Vendor: 0428 Product: 4001)",
         "Linux",
         GravisGamepadProSchema),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "0428-4001-Gravis GamePad Pro USB ",
         "Linux",
         GravisGamepadProSchema),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "GamePad Pro USB  (Vendor: 0428 Product: 4001)",
         "macOS",
         GravisGamepadProSchema),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "generic",
         "428-4001-GamePad Pro USB ",
         "macOS",
         GravisGamepadProSchema),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "xboxone",
         "Performance Designed Products Afterglow Wired Controller for Xbox One (STANDARD GAMEPAD Vendor: 0e6f Product: 0139)",
-        "Linux",
-        GamepadAPI.Stardard
+        "Linux"
     ),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "xboxone",
         "Performance Designed Products Afterglow Wired Controller for Xbox One (Vendor: 0e6f Product: 0139)",
         "Linux",
@@ -891,7 +892,7 @@ const GameInputModels = [
             /* r_trigger */  new SchemaAxisButton(6)
         )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "xboxone",
         "0e6f-0139-Afterglow Prismatic Wired Controller",
         "Linux",
@@ -919,7 +920,7 @@ const GameInputModels = [
             /* r_trigger */  new SchemaAxisButton(6)
         )),
     new GameInputModel(
-        GameInputSchemaType.Plumber,
+        GameInputSchema.Plumber,
         "nintendo-generic",
         "USB,2-axis 8-button gamepad   (STANDARD GAMEPAD Vendor: 0583 Product: 2060)",
         "macOS",
@@ -947,7 +948,7 @@ const GameInputModels = [
             /* r_trigger */  undefined
         )),
     new GameInputModel(
-        GameInputSchemaType.Plumber,
+        GameInputSchema.Plumber,
         "nintendo-generic",
         "0583-2060-USB,2-axis 8-button gamepad   ",
         "macOS",
@@ -975,7 +976,7 @@ const GameInputModels = [
             /* r_trigger */  undefined
         )),
     new GameInputModel(
-        GameInputSchemaType.Plumber, // it's a square and x /y are generic style, while a / b are nintendo style - FAMICOM
+        GameInputSchema.Plumber, // it's a square and x /y are generic style, while a / b are nintendo style - FAMICOM
         "nintendo-generic",
         "BUFFALO BGC-FC801 USB Gamepad (Vendor: 0411 Product: 00c6)",
         "macOS",
@@ -1003,7 +1004,7 @@ const GameInputModels = [
             /* r_trigger */  undefined
         )),
     new GameInputModel(
-        GameInputSchemaType.Plumber, // it's a square and x /y are generic style, while a / b are nintendo style - FAMICOM
+        GameInputSchema.Plumber, // it's a square and x /y are generic style, while a / b are nintendo style - FAMICOM
         "nintendo-generic",
         "0411-00c6-BUFFALO BGC-FC801 USB Gamepad ",
         "macOS",
@@ -1031,7 +1032,7 @@ const GameInputModels = [
             /* r_trigger */  undefined
         )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog, // should be numbers, see issue #16
+        GameInputSchema.Hedgehog, // should be numbers, see issue #16
         "generic",
         "USB  Joystick (Vendor: 1345 Product: 1030)",
         "macOS",
@@ -1059,7 +1060,7 @@ const GameInputModels = [
             /* r_trigger */  8
         )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog, // should be numbers, see issue #16
+        GameInputSchema.Hedgehog, // should be numbers, see issue #16
         "generic",
         "1345-1030-USB  Joystick ",
         "macOS",
@@ -1087,7 +1088,7 @@ const GameInputModels = [
             /* r_trigger */  8
         )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "joystick",
         "Innovative Produtcs VR Pro 2000 USB Joystick (Vendor: 04b4 Product: 2774)",
         "Linux",
@@ -1115,7 +1116,7 @@ const GameInputModels = [
             /* r_trigger */  8
         )),
     new GameInputModel(
-        GameInputSchemaType.Hedgehog,
+        GameInputSchema.Hedgehog,
         "joystick",
         "04b4-2774-Innovative Produtcs VR Pro 2000 USB Joystick",
         "Linux",
@@ -1143,7 +1144,7 @@ const GameInputModels = [
             /* r_trigger */  8
         )),
     new GameInputModel(
-        GameInputSchemaType.Plumber,
+        GameInputSchema.Plumber,
         "nintendo-generic",
         "20d6-a711-Bensussen Deutsch & Associates,Inc.(BDA) Core (Plus) Wired Controller",
         "Linux",
@@ -1172,7 +1173,7 @@ const GameInputModels = [
         )
     ),
     new GameInputModel(
-        GameInputSchemaType.Plumber,
+        GameInputSchema.Plumber,
         "nintendo-generic",
         "Bensussen Deutsch & Associates,Inc.(BDA) Core (Plus) Wired Controller (Vendor: 20d6 Product: a711)",
         "Linux",
@@ -1201,7 +1202,7 @@ const GameInputModels = [
         )
     ),
     new GameInputModel(
-        GameInputSchemaType.Plumber,
+        GameInputSchema.Plumber,
         "nintendo-generic",
         "Core (Plus) Wired Controller (Vendor: 20d6 Product: a711)",
         "Windows",
