+++ src/gamepad-schema-names.js
@@ -1,0 +1,37 @@
+/**
+ * @preserve
+ * @source: https://raw.githubusercontent.com/lunarcloud/gameinputjs/master/gamepad-schema-names.js
+ * @license magnet:?xt=urn:btih:d3d9a9a6595521f9666a5e94cc830dab83b65699&dn=expat.txt MIT (Expat) License
+ */
+
+const GamepadSchemaNames = {
+    d_up        :   "d_up",
+    d_down      :   "d_down",
+    d_left      :   "d_left",
+    d_right     :   "d_right",
+    menu        :   "menu",
+    button0     :   "button0",
+    button1     :   "button1",
+    button2     :   "button2",
+    button3     :   "button3",
+    l_up        :   "l_up",
+    l_down      :   "l_down",
+    l_left      :   "l_left",
+    l_right     :   "l_right",
+    r_up        :   "r_up",
+    r_down      :   "r_down",
+    r_left      :   "r_left",
+    r_right     :   "r_right",
+    l_button    :   "l_button",
+    r_button    :   "r_button",
+    l_trigger   :   "l_trigger",
+    r_trigger   :   "r_trigger"
+}
+
+export default GamepadSchemaNames
+export {GamepadSchemaNames}
+
+/**
+ * @preserve
+ * @license-end
+ */
