+++ src/tester.html
@@ -9,6 +9,7 @@
     <link rel="stylesheet"href="https://fonts.googleapis.com/css?family=Raleway:400,300,600">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/4.1.1/normalize.min.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.min.css">
+    <link rel="stylesheet" href="css/tester-common.css">
     <style>
         .template {
             display: none !important;
@@ -127,12 +128,61 @@
 <!-- End Document
   –––––––––––––––––––––––––––––––––––––––––––––––––– -->
 <script type="module">
-    import { Vector2, GamepadAPI, GameInputModel, GameInputSchemaType, SchemaAxisButton } from './gameinput.types.js'
-    import { GameInputDom, GameInput } from './gameinput.dom.js'
+    import { GameInput, GamepadSchemaNames } from './gameinput.js'
 
     let gi = new GameInput();
     gi.debug = true;
-    GameInputDom.setupDom(gi);
+
+    // Setup
+    gi.onButtonDown((playerIndex, schemaName) => {
+        var buttonElements = document.querySelectorAll(".gameinput-player" + playerIndex + "-" + schemaName);
+        for (var i = 0; i < buttonElements.length; i++) {
+            buttonElements[i].classList.add("gameinput-button-active");
+        }
+    });
+
+    gi.onButtonUp((playerIndex, schemaName) => {
+        var buttonElements = document.querySelectorAll(".gameinput-player" + playerIndex + "-" + schemaName);
+        for (var i = 0; i < buttonElements.length; i++) {
+            buttonElements[i].classList.remove("gameinput-button-active");
+        }
+    });
+
+    gi.onReshufflePlayers(function()
+    {
+        /* Remove Old Stuff */
+        for (var i = 0; i < gi.Players.length; i++)
+        {
+            var previousThemeStyleElements = document.head.querySelectorAll('.gameinput-theme-player' + i);
+            for (let j = 0; j < previousThemeStyleElements.length; j++) document.head.removeChild(previousThemeStyleElements[j]);
+        }
+
+        setTimeout(function(){
+            /* Add new stuff */
+            for (var i = 0; i < gi.Players.length; i++)
+            {
+                var player = gi.getPlayer(i);
+
+                if (typeof(player.type) !== "undefined" && typeof(player.model) !== "undefined") {
+                    var playerIcons = document.querySelectorAll("img.gameinput-icon-player" + player.index);
+                    for (let j = 0; j < playerIcons.length; j++ ) playerIcons[j].setAttribute("src", "img/" + player.model.iconName + ".png");
+
+                    var backgroundIcons = document.querySelectorAll(".gameinput-icon-background-player" + player.index);
+                    for (let j = 0; j < backgroundIcons.length; j++ ) backgroundIcons[j].style.backgroundImage = "url('img/models/" + player.model.iconName + ".png')";
+
+                    var themeStyleElement = document.createElement('link');
+                    themeStyleElement.classList.add("gameinput-theme-player" + player.index);
+                    themeStyleElement.setAttribute("rel", "stylesheet");
+                    themeStyleElement.setAttribute("href", "css/" + player.type.theme.name.toLowerCase() + "/" + player.index + ".css");
+                    document.head.appendChild(themeStyleElement);
+                } else {
+                    var previousPlayerIcons = document.querySelectorAll("img.gameinput-icon-player" + player.index);
+                    for (let j = 0; j < previousPlayerIcons.length; j++ ) previousPlayerIcons[j].setAttribute("src", "");
+                }
+            }
+        },1);
+    });
+    
     gi.onReshufflePlayers(() => setupPlayers());
     setTimeout(()=> gi.initialGamePadSetup(), 100); // TODO?
 
@@ -165,7 +215,7 @@
                 iconPlayerReferences[k].classList.remove("gameinput-icon-player-template");
             }
 
-            for (var j in GameInput.SchemaNames)
+            for (var j in GamepadSchemaNames)
             {
                 var templateUpReferences = row.querySelectorAll(".gameinput-player-template-up");
                 for (var k = 0; k < templateUpReferences.length; k++) {
@@ -188,19 +238,19 @@
                     templateRightReferences[k].classList.remove("gameinput-player-template-right");
                 }
 
-                if (typeof(gi.getPlayer(i).schema[GameInput.SchemaNames[j]]) === "undefined")
+                if (typeof(gi.getPlayer(i).schema[GamepadSchemaNames[j]]) === "undefined")
                 {
-                    var unmappedReferences = row.querySelectorAll(".gameinput-player-template-" + GameInput.SchemaNames[j]);
+                    var unmappedReferences = row.querySelectorAll(".gameinput-player-template-" + GamepadSchemaNames[j]);
                     for (var k = 0; k < unmappedReferences.length; k++) {
                         unmappedReferences[k].parentElement.removeChild(unmappedReferences[k]);
                     }
                 }
                 else
                 {
-                    var templateReferences = row.querySelectorAll(".gameinput-player-template-" + GameInput.SchemaNames[j]);
+                    var templateReferences = row.querySelectorAll(".gameinput-player-template-" + GamepadSchemaNames[j]);
                     for (var k = 0; k < templateReferences.length; k++) {
-                        templateReferences[k].classList.add("gameinput-player" + i + "-" + GameInput.SchemaNames[j]);
-                        templateReferences[k].classList.remove("gameinput-player-template-" + GameInput.SchemaNames[j]);
+                        templateReferences[k].classList.add("gameinput-player" + i + "-" + GamepadSchemaNames[j]);
+                        templateReferences[k].classList.remove("gameinput-player-template-" + GamepadSchemaNames[j]);
                     }
                 }
             }
